+++ bin/structure_building_cli.dart
@@ -1,11 +1,12 @@
-import 'dart:io';
-
-import 'package:structure_building_cli/structure_building_cli.dart';
+import 'package:structure_building_cli/command/command_getter.dart';
 
 Future<void> main(List<String> arguments) async {
-  stdout.write('Please enter your screen name:');
-  String? screenName = stdin.readLineSync();
+  var time = Stopwatch();
+  time.start();
 
-  if (screenName == null) stderr.write('Screen name is null');
-  ScreenBuilder.create(screenName!);
+  final command = CommandGetter.find(arguments);
+  await command.execute();
+
+  time.stop();
+  print('Time: ${time.elapsed.inMilliseconds} Milliseconds');
 }