+++ lib/command/icommand.dart
@@ -1,0 +1,7 @@
+abstract class ICommand {
+  String get command;
+
+  String get errorMessage;
+
+  Future<void> execute();
+}