+++ lib/structure_building_cli.dart
@@ -1,1 +1,2 @@
 export 'screen_builder/screen_builder.dart';
+export 'project_builder/project_builder.dart';