+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.4
+
+- project create cli change.
+
+## 1.0.3
+
+- Change app name and bundle id cli added.
+
+## 1.0.2
+
+- Create Project added.
+
+## 1.0.1
+
+- Executables added.
+
 ## 1.0.0
 
 - Initial version.